Review changes: used FileUIActivity, fixed logs, PUT instead of POST
