+++ core/connector/src/main/java/org/exoplatform/wcm/connector/collaboration/DocumentEditorsRESTService.java
@@ -26,6 +26,7 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.FormParam;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -170,7 +171,7 @@ public class DocumentEditorsRESTService implements ResourceContainer {
    * @param documentEditorData the editor provider DTO
    * @return the response
    */
-  @POST
+  @PUT
   @Path("/{provider}")
   @RolesAllowed("administrators")
   @Produces(MediaType.APPLICATION_JSON)
